pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "workout-tracker"
version = "0.1.0"
description = "Encrypted workout tracking service with a FastAPI backend and React frontend."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "Workout Tracker Authors" }
]
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.29",
  "pydantic>=2.7.1",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "cryptography>=42.0.5",
  "passlib>=1.7.4",
  "webauthn>=2.1.0",
  "pyjwt[crypto]>=2.8.0",
  "httpx>=0.27.0",
  "email-validator>=2.1.1",
  "python-multipart>=0.0.9",
  "itsdangerous>=2.2.0",
  "psycopg[binary,pool]>=3.1.18",
]

[project.optional-dependencies]
postgres = ["psycopg[binary,pool]>=3.1.18"]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
workout-tracker-api = "workout_tracker.app:run"

[tool.hatch.build.targets.wheel]
packages = ["src/workout_tracker"]
force-include = { "frontend/dist" = "workout_tracker/static" }

[tool.hatch.metadata]
allow-direct-references = true
